+++ app/mailers/matilda_core/application_mailer.rb
@@ -6,7 +6,7 @@ module MatildaCore
   class ApplicationMailer < ActionMailer::Base
 
     layout 'matilda_core/mailer'
-    default from: 'Matilda <matilda@1d3o.it>'
+    default from: MatildaCore.config.mailer_from_address
     default template_path: 'matilda_core/mailers'
 
   end